bin/requirements.txt
# python 3.12
biopython==1.85
Jinja2==3.1.5
numpy==2.2.3
pandas==2.2.3
pyyaml==6.0.2
